# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Gestione Utenti",
    "description": "Documentazione per le API utilizzate per il sito Marvel"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "Users",
      "description": "Gestione degli utenti"
    },
    {
      "name": "Auth",
      "description": "Autenticazione e Login"
    },
    {
      "name": "Credits",
      "description": "Gestione dei crediti utente"
    },
    {
      "name": "Trade",
      "description": "Gestione degli scambi delle carte"
    },
    {
      "name": "Marvel",
      "description": "Utilizzo delle API della Marvel"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/users/findAll": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Ottieni tutti gli utenti",
        "description": "Restituisce una lista di tutti gli utenti registrati.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/add": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Aggiungi un nuovo utente",
        "description": "Aggiunge un utente con i dati forniti.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                },
                "name": {
                  "example": "any"
                },
                "surname": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "fav_hero": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/users/update": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Modifica un utente",
        "description": "Modifica un utente con i dati forniti.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "old_username": {
                  "example": "any"
                },
                "username": {
                  "example": "any"
                },
                "name": {
                  "example": "any"
                },
                "surname": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "fav_hero": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/find": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Trova un utente",
        "description": "Trova un utente attraverso l",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/delete": {
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Elimina un utente",
        "description": "Elimina un utente fornendo l",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Effettua il login",
        "description": "Verifica le credenziali e permette l",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/auth": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Verifica autenticazione",
        "description": "Verifica se l",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/updatecredits": {
      "post": {
        "tags": [
          "Credits"
        ],
        "summary": "Aggiorna i crediti di un utente",
        "description": "Permette di aggiornare il numero di crediti di un utente.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "credits": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/add": {
      "post": {
        "tags": [
          "Trade"
        ],
        "summary": "Crea una offerta per lo scambio",
        "description": "Verifica che le carte da scambiare non siano già impiegate in altri scambi e poi crea l",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "cards": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/show": {
      "get": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/marvel/marvellous": {
      "post": {
        "tags": [
          "Marvel"
        ],
        "summary": "Effettua una richiesta sulle API Marvel",
        "description": "Utilizza una query per ricercare le informazioni messe a disposizione dalla Marvel",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "url": {
                  "example": "any"
                },
                "query": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/marvel/unpack": {
      "get": {
        "tags": [
          "Marvel"
        ],
        "summary": "Apre un pacchetto di 5 carte",
        "description": "Fornisce a un utente autenticato 5 carte casuali da aggiungere alla collezione",
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Users": {
        "$username": "Superhero1",
        "$name": "David",
        "$surname": "Pizzolato",
        "$password": "PwMoltoBella12345",
        "$email": "david.pizzolato1@studenti.unimi.it",
        "$fav_hero": "Superman",
        "$credits": "7040"
      }
    }
  }
}